requirements.txt
aiogram>=3.15.0
httpx>=0.28.0
python-dotenv>=1.0.0
